feat:convert sold cars to column for parquet
